improve multicluster support
Adapting the script to run with the mstart.
I introduced a new variable called `NUMBER_OF_CLUSTERS`. All the
clusters deployed are on the same container but the dashboard is
accessbile on different ports.

The first port is configurable and a 100 will be added to the ports of
next cluster.

for eg. CEPH1 is in 11000 and CEPH2 is in 11100 and CEPH3 is in
11200....

LIMITATIONS SO FAR
DEV SERVER ON FIRST CLUSTER ONLY
localhost:11000 is somehow not there :(
MONITORING STACKS CAN ONLY DEPLOYED ON FIRST CLUSTER

Signed-off-by: Nizamudeen A <nia@redhat.com>