Update docker-compose.yml
This update replaces the deprecated jboss/keycloak:11.0.3 image in the docker-compose.yml file with the newer and supported quay.io/keycloak/keycloak:21.1.1 image. The configuration has been updated to align with the requirements of the newer Keycloak versions. This includes:

1) Changing KEYCLOAK_USER and KEYCLOAK_PASSWORD to KEYCLOAK_ADMIN and KEYCLOAK_ADMIN_PASSWORD.
2) Adding the start --hostname keycloak command for compatibility with the new Keycloak startup process.
3) Ensuring the file remains functional for all other services.
These updates ensure compatibility with the latest Keycloak releases and address the deprecation of the old image.